Extend Analytics adapter (#13)
Add the concept of bidWinner. Before this change, I was using isTop to
identify the winner of the impression. After this change, isTop is true
when that partner won the auction before going to the Ad Server and
isWinner is true when the partner won the auction and won in the AdServer.