SOR-2753: Use better way to get href of page
According to comment in Prebid, getTopWindowLocation is being deprecated
and we should use the refererInfo object instead. If that doesn't exist,
go back to using gTWL.

This should fix the issue for standalone customers, but won't take
effect until this change is merged back into the Prebid master branch
and customers update.

Container fix will be done separately/differently, since that version
doesn't appear to have access to refererInfo.